Use RocksDbMemoizationStore
