+++ Public/Src/Cache/MemoizationStore/Library/BuildXL.Cache.MemoizationStore.Library.dsc
@@ -18,8 +18,6 @@ namespace Library {
             ContentStore.Library.dll,
             Interfaces.dll,
             
-            importFrom("BuildXL.Cache.DistributedCache.Host").Configuration.dll,
-            importFrom("BuildXL.Cache.DistributedCache.Host").Service.dll,
             importFrom("BuildXL.Utilities").dll,
 
             importFrom("System.Data.SQLite.Core").pkg,
